Close V2 history and token-auth client gaps
Protocol V2 needed client-facing continuity and auth surfaces without breaking existing Pusher HMAC behavior. This change extends the existing channel history action, captures attach serials for untilAttach reads, adds HS256 capability-token auth and revocation, and wires token capabilities into the existing subscription, history, publish, presence, and mutation checks.

Constraint: Preserve V1 Pusher-compatible HMAC and pusher:signin behavior

Constraint: Reuse existing HistoryStore, rewind, recovery, ConnectionCapabilities, and app-key HTTP auth paths

Rejected: Add a parallel sockudo:history frame | Existing SDKs already send channel_history and the response contract must remain additive

Rejected: Replace signin/HMAC capabilities with JWT-only auth | Existing clients depend on the current auth model

Confidence: medium

Scope-risk: broad

Directive: Do not loosen token-auth mutation checks without verifying publish capability and existing own/any policies together

Tested: cargo fmt --all

Tested: cargo test -p sockudo-core capability_token::tests

Tested: cargo check -p sockudo-adapter --all-targets

Tested: cargo test -p sockudo-adapter --all-targets

Tested: cargo check -p sockudo --all-targets

Tested: cargo test -p sockudo --all-targets

Not-tested: fuzz targets and performance benches

+++ crates/sockudo-adapter/src/handler/core.rs
@@ -38,7 +38,7 @@ impl ConnectionHandler {
     ) -> Result<()> {
         let error_data = ErrorData {
             message: error.to_string(),
-            code: Some(error.close_code()),
+            code: Some(u32::from(error.close_code())),
         };
         let error_message =
             PusherMessage::error(error_data.code.unwrap_or(4000), error_data.message, channel);
